🔨 Restructure components
