+++ src/components/cards/ListCard.vue
